+++ pkg/clients/grpc/manager/manager_test.go
@@ -7,19 +7,19 @@ import (
 
 	"github.com/absmach/supermq/pkg/errors"
 	"github.com/stretchr/testify/assert"
-	"github.com/ultravioletrs/cocos/pkg/clients/grpc"
+	"github.com/ultravioletrs/cocos/pkg/clients"
 )
 
 func TestNewManagerClient(t *testing.T) {
 	tests := []struct {
 		name string
-		cfg  grpc.ManagerClientConfig
+		cfg  clients.StandardClientConfig
 		err  error
 	}{
 		{
 			name: "Valid config",
-			cfg: grpc.ManagerClientConfig{
-				BaseConfig: grpc.BaseConfig{
+			cfg: clients.StandardClientConfig{
+				BaseConfig: clients.BaseConfig{
 					URL: "localhost:7001",
 				},
 			},
